Former-commit-id: bcd0eeb753d62d34cbb301b460b6e0c49f2dafa6

+++ rs-dataprovider/acquisition/acquisition-plugins-ssalto/src/test/java/fr/cnes/regards/modules/acquisition/plugins/ssalto/chain/Jason2CorpfaltiPoeProcessingChainTest.java
@@ -120,12 +120,12 @@ public class Jason2CorpfaltiPoeProcessingChainTest extends AbstractAcquisitionCh
     }
 
     @Override
-    protected int getExpectedFiles() {
+    protected long getExpectedFiles() {
         return 1;
     }
 
     @Override
-    protected int getExpectedProducts() {
+    protected long getExpectedProducts() {
         return 1;
     }
