+++ rs-dataprovider/acquisition/acquisition-plugins-ssalto/src/test/java/fr/cnes/regards/modules/acquisition/plugins/ssalto/chain/Spot2Doris1bProcessingChainTest.java
@@ -131,12 +131,12 @@ public class Spot2Doris1bProcessingChainTest extends AbstractAcquisitionChainTes
     }
 
     @Override
-    protected int getExpectedFiles() {
+    protected long getExpectedFiles() {
         return 1;
     }
 
     @Override
-    protected int getExpectedProducts() {
+    protected long getExpectedProducts() {
         return 1;
     }
 }
